consolidate all layout wrappers to layout component
