+++ lib/testUtils.tsx
@@ -3,48 +3,21 @@ import userEvent from '@testing-library/user-event'
 import { StatusCodes } from 'http-status-codes'
 import { delay, http, HttpResponse, type JsonBodyType, type Path } from 'msw'
 import { type NextApiHandler } from 'next'
-import { SessionProvider } from 'next-auth/react'
+import { type Session } from 'next-auth'
 import {
   type NtarhInitPagesRouter,
   testApiHandler,
 } from 'next-test-api-route-handler'
-import { type ReactElement, type ReactNode } from 'react'
-import { SWRConfig } from 'swr'
+import { type ReactElement } from 'react'
+import { expect, vi } from 'vitest'
+import Layout from '../components/Layout'
 import { type ApiError } from '../models/ApiError'
 import { server } from '../msw-mocks/server'
 import { methodNotAllowed } from './backend/apiMiddleware/util'
 import { devUserId } from './frontend/constants'
-import { type Session } from 'next-auth'
-import { vi, expect } from 'vitest'
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider'
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 
 // This file overwrites @testing-library's render and wraps it with components that
 // need to be set up for every test.
-
-// Note: fetch() needs to be polyfilled or it will be undefined (just need to add "import 'whatwg-fetch'" in the test setup file).
-// See: https://testing-library.com/docs/react-testing-library/setup/#configuring-jest-with-test-utils
-// See: https://mswjs.io/docs/faq#swr
-const FrontendLayout = ({
-  children,
-  user,
-}: {
-  children: ReactNode
-  user?: Session['user']
-}) => (
-  <SWRConfig
-    value={{
-      fetcher: (url: string) => fetch(url).then((r) => r.json()),
-      provider: () => new Map(),
-    }}
-  >
-    <SessionProvider session={{ user: user ?? { id: devUserId }, expires: '' }}>
-      <LocalizationProvider dateAdapter={AdapterDayjs}>
-        {children}
-      </LocalizationProvider>
-    </SessionProvider>
-  </SWRConfig>
-)
 
 /** Custom render implementation that wraps the element with necessary layout components (eg, SWRConfig).
  *  Also returns a user object to remove the need to setup a userEvent.
@@ -75,7 +48,18 @@ const customRender = (
   ),
   ...render(ui, {
     wrapper: ({ children }) =>
-      FrontendLayout({ children, user: options?.user }),
+      Layout({
+        children,
+        disableNavbar: true,
+        swrConfig: {
+          // Note: fetch() needs to be polyfilled or it will be undefined (just need to add "import 'whatwg-fetch'" in the test setup file).
+          // See: https://testing-library.com/docs/react-testing-library/setup/#configuring-jest-with-test-utils
+          // See: https://mswjs.io/docs/faq#swr
+          fetcher: (url: string) => fetch(url).then((r) => r.json()),
+          provider: () => new Map(),
+        },
+        session: { user: options?.user ?? { id: devUserId }, expires: '' },
+      }),
     ...options,
   }),
 })
